requirement.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
django==5.0.1; python_version >= '3.10'
djangorestframework==3.14.0; python_version >= '3.6'
pytz==2023.4
sqlparse==0.4.4; python_version >= '3.5'
